Fix behavior of golangci-lint against Go comments
These rules weren't not enabled in golangci-lint v1.
Because we enabled EXC0014, the migration tool disabled the "comment" presets.

So it removed all the exclusions about the Go comments.
While we only expected to remove the EXC0014 exclusion.

This commit restores the previous and expected behavior.
We might want to re-enable them in the future, but for now.