Migrate to golangci-lint v2 (#1708)
* Use golangci-lint migration tool

I simply launched the following command

$ golangci-lint migrate

The configuration file was updated to the new format.

The comments were stripped by the migration tool.

* Restore the comments stripped by golangci-lint migration tool

I copy pasted them manually from the v1 configuration file.

* Remove useless exclusions from golangci-lint

Also fix "examples" folder exclusions.

This folder was previously excluded by golangci-lint v1.
The migration tool failed to migrate this.

Later, we can remove the exclusions for "examples".
But the idea is to keep the same behavior as before.

* Remove legacy "lax" option from golangci-lint config

We don't need it, we can use the default value.

lax was the default behavior in the golangci-lint v1.
We can remove it as we don't need it.

* Add TODO comments for golangci-lint presets

We may want to fix these errors reported by golangci-lint in the future.
But as they were already in golangci-lint v1, we will not fix them now.

* Remove some staticcheck rules from golangci-lint config

These rules were not activated in golangci-lint v1.
They were added in v2 by default, but they are reporting a lot of issues.

This commit restores the previous behavior.
We might want to re-enable them in the future, but for now.

* Fix behavior of golangci-lint against Go comments

These rules weren't not enabled in golangci-lint v1.
Because we enabled EXC0014, the migration tool disabled the "comment" presets.

So it removed all the exclusions about the Go comments.
While we only expected to remove the EXC0014 exclusion.

This commit restores the previous and expected behavior.
We might want to re-enable them in the future, but for now.

* Apply suggested fixes provided by latest testifylint version

testifylint was improved, so it reports more issues now.

These errors were fixed by simply running the following command:

$ golangci-lint run --fix --enable-only testifylint

* Fix errors reported by testifylint

pkg/gofr/datasource/arangodb/arango_document_test.go:51:2: formatter: using msgAndArgs with non-string first element (msg) causes panic (testifylint)
	require.ErrorIs(t, err, errDocumentNotFound, err, "Expected error when document not found")
	^
pkg/gofr/datasource/arangodb/arango_document_test.go:87:2: formatter: using msgAndArgs with non-string first element (msg) causes panic (testifylint)
	require.ErrorIs(t, err, errDocumentNotFound, err, "Expected error when document not found")

* Apply suggested fixes provided by latest gci version

gci was improved, so it reports more issues now.

These errors were fixed by simply running the following command:

$ golangci-lint fmt --fix

* Migrate GitHub Actions to use golangci-lint v2

---------

Co-authored-by: Umang Mundhra <mundhraumang.02@gmail.com>